Refactor
Moved the task for handling the creation of a `ScheduledTask` into the `ScheduledTskService` . Added a enum for the `TaskType`.
