+++ Data/Enums/TaskType.cs
@@ -1,0 +1,5 @@
+public enum TaskType
+{
+    TestTask,
+
+}
